Move page rendering logic to pages module
