+++ src/csrs/pages/edit/__init__.py
@@ -1,0 +1,1 @@
+from . import render
