+++ src/csrs/routes/forms.py
@@ -5,96 +5,9 @@ from sqlalchemy.orm import Session
 from .. import crud, errors
 from ..database import get_db
 from ..logger import logger
-from ..pages import loader, templates
+from ..pages.edit import render
 
-ALLOW_EDITING_VIA_FORMS = True
 router = APIRouter(prefix="/forms", include_in_schema=False)
-
-
-###############################################################################
-# RENDER
-# Below are page rendering functions
-
-
-def render_assumptions(request: Request, db: Session):
-    all_objs = crud.assumptions.read(db=db)
-    all_kinds = crud.assumptions.read_kinds(db=db)
-    objects = [templates.EditableAssumption(obj, all_kinds) for obj in all_objs]
-    metadata = loader.ENV.get_template("static/metadata/assumption.jinja").render()
-    return loader.jinja_loader.TemplateResponse(
-        "templates/edit.jinja",
-        {
-            "request": request,
-            "page_title": "Assumptions",
-            "metadata": metadata,
-            "objects": objects,
-            "new_object": templates.NewAssumption(all_kinds),
-            "edit_on": ALLOW_EDITING_VIA_FORMS,
-        },
-    )
-
-
-def render_scenarios(request: Request, db: Session):
-    all_objs = crud.scenarios.read(db=db)
-    all_assumptions = crud.assumptions.read(db=db)
-    objects = list()
-    for obj in all_objs:
-        versions = [r.version for r in crud.runs.read(db=db, scenario=obj.name)]
-        t = templates.EditableScenario(obj, versions, all_assumptions)
-        objects.append(t)
-    metadata = loader.ENV.get_template("static/metadata/scenario.jinja").render()
-    return loader.jinja_loader.TemplateResponse(
-        "templates/edit.jinja",
-        {
-            "request": request,
-            "page_title": "Scenarios",
-            "metadata": metadata,
-            "objects": objects,
-            "new_object": templates.NewScenario(all_assumptions),
-            "edit_on": ALLOW_EDITING_VIA_FORMS,
-        },
-    )
-
-
-def render_runs(request: Request, db: Session):
-    all_objs = crud.runs.read(db=db)
-    all_scenarios = crud.scenarios.read(db=db)
-    objects = list()
-    for obj in all_objs:
-        t = templates.EditableRuns(obj)
-        objects.append(t)
-    metadata = loader.ENV.get_template("static/metadata/run.jinja").render()
-    return loader.jinja_loader.TemplateResponse(
-        "templates/edit.jinja",
-        {
-            "request": request,
-            "page_title": "Runs",
-            "metadata": metadata,
-            "objects": objects,
-            "new_object": templates.NewRuns(all_scenarios),
-            "edit_on": ALLOW_EDITING_VIA_FORMS,
-        },
-    )
-
-
-def render_paths(request: Request, db: Session):
-    all_objs = crud.paths.read(db=db)
-    objects = list()
-    for obj in all_objs:
-        t = templates.EditablePaths(obj)
-        objects.append(t)
-    metadata = loader.ENV.get_template("static/metadata/path.jinja").render()
-    return loader.jinja_loader.TemplateResponse(
-        "templates/edit.jinja",
-        {
-            "request": request,
-            "page_title": "Named Paths",
-            "metadata": metadata,
-            "objects": objects,
-            "new_object": templates.NewPath(),
-            "edit_on": ALLOW_EDITING_VIA_FORMS,
-        },
-    )
 
 
 ###############################################################################
@@ -105,25 +18,25 @@ def render_paths(request: Request, db: Session):
 @router.get("/assumptions", response_class=HTMLResponse)
 async def form_assumptions(request: Request, db: Session = Depends(get_db)):
     logger.info(f"{request.method} {request.url}")
-    return render_assumptions(request, db)
+    return render.render_assumptions(request, db)
 
 
 @router.get("/scenarios", response_class=HTMLResponse)
 async def form_scenarios(request: Request, db: Session = Depends(get_db)):
     logger.info(f"{request.method} {request.url}")
-    return render_scenarios(request, db)
+    return render.render_scenarios(request, db)
 
 
 @router.get("/runs", response_class=HTMLResponse)
 async def form_runs(request: Request, db: Session = Depends(get_db)):
     logger.info(f"{request.method} {request.url}")
-    return render_runs(request, db)
+    return render.render_runs(request, db)
 
 
 @router.get("/paths", response_class=HTMLResponse)
 async def form_paths(request: Request, db: Session = Depends(get_db)):
     logger.info(f"{request.method} {request.url}")
-    return render_paths(request, db)
+    return render.render_paths(request, db)
 
 
 ###############################################################################
@@ -438,7 +351,7 @@ async def form_paths_delete(
     return RedirectResponse(request.url_for("form_paths"), status_code=302)
 
 
-if ALLOW_EDITING_VIA_FORMS:
+if render.ALLOW_EDITING_VIA_FORMS:
     # Assumptions
     router.post(
         "/assumptions/create",
